Update job-scheduler.md
Coming newly into the orchestration team, I am updating the initial description of the scheduler types to better align with how users may view the product today: 1. we refer to merge jobs as merge jobs (and it is an explicit category in the UI) so I think it would be useful to call out that term in the beginning, so people associate that type of job with its name and 2. another event-based job is job completion (another explicit category in the UI) and I couldn't tell how that fit into the initial description, so added that as well. Feel free to revise, just wanted to flag those two points!